Fixed unexpected filter behaviour with favorites
The filter was checking the "highlighted" property of the talks to determine whether it is a favorite or not. Unfortunately this property is only used to display a "highlight" indicator in master/detail mode, and gives no information about whether the talk is a favorite or not.

The result was that was recently selected to automatically became a favorite, and none of the real favorites ever showed up. Fixed by making the TalkListAdapter use the "Favorites" manager class instead of the highlight property.