Abandoned experiment: tabular layout.
This is a bit of code to work towards a table layout where rows are time slots
and columns are rooms or tracks. It explores the concept and it proves that all
the information we need to build this dynamically is there, but there's quite a
bit of work to make it happen:

* Determine number of rows:
  -> Check for the number of unique Talk.getSlot();
* Determine the number of columns:
  -> Collect number of talks by slot (TalkListAdapter L131-144)
* Work out which talks are full-width
  -> A simple heuristic would be to look for slots with only one talk.
* Build an Adapter that offers the talks in tabular form to the
  RecyclerView.LayoutManager
* Create an alternative talk card layout that's narrow and a bit taller.
* Update the talk list fragment to wire it all together.
* Update the talk list activity to make the new view accessible.

For the time being, a shorter path to our goals is to just insert the time
slots as section headers in the main talk list. We may revisit this
multi-column thing later.