Proves concept of tag colors in filter menu.
This uses DrawableCompat.setTint() to put Drawable tinting to work even on
older platform versions (tested on an API 15 emulator). It adds an icon to each
of the tag-filter menu items and renders that icon in the Chosen color for that
tag.

Sticking an icon on a menu item that's already got a radio button on its other
end is crude and awful. We should do something much more subtle, such as a
little vertical band like we have on the card views themselves. This means
setting a background on the menu item, which means creating an ActionProvider
to produce menu widgets. I can do that, but I don't have the time to do so
right this instant, hence, this concept-proving commit, which you probably
shouldn't merge into master.