Makes for narrower color bands in the filter menu.
The preceding commit puts an icon to the left of every filter menu item, taking
up a lot of space. This commit uses a text span to add a margin to the left of
every filter menu item, with a colored strip rendered into it. This takes up
less space than the icons.